game and gameVariant changes on controller and service

+++ wbook-api/Infrastructure/Models/Game/GameVariantsModel.cs
@@ -15,8 +15,8 @@ namespace Infrastructure.Models.Game
         public string Description { get; set; }
         public string Image { get; set; }
         public string? HoverImage { get; set; }
-        public int Price { get; set; }
-        public int Rrp { get; set; }
+        public double Price { get; set; }
+        public double Rrp { get; set; }
         public List<GameVariantItemModel> GameVariants { get; set; }
     }
 }
